refactor and fix typos
