--- common/errror.go
@@ -1,7 +1,0 @@
-package common
-
-import (
-	"errors"
-)
-
-var ErrRecordNotFound = errors.New("record not found")
